+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Added a new `ExpireCacheService` class that handles marking cache as expired when the refresh mode is set to expire the cache.
-- Added a new `cacheControlHeaderExpired` config setting that reduces the max cache age for public reverse proxies to 5 seconds by default for expired pages.
+- Added a new `cacheControlHeaderExpired` config setting that reduces the max cache age in public reverse proxies to 5 seconds by default for expired pages.
 
 ### Changed